pyproject.toml
[project]
name = "trading-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pycryptodome>=3.23.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]
